refactor: theme section 분리
